Reorganized code Added a viewmodel for the MainActivity
