+++ app/src/main/java/com/amalitech/bookmeetingroom/BookMeetingApplication.kt
@@ -1,6 +1,7 @@
 package com.amalitech.bookmeetingroom
 
 import android.app.Application
+import com.amalitech.bookmeetingroom.di.appModule
 import com.amalitech.core.di.coreModule
 import com.amalitech.core.di.networkModule
 import com.amalitech.onboarding.di.onboardingDomainModule
@@ -19,7 +20,8 @@ class BookMeetingApplication : Application() {
                 networkModule,
                 onboardingDataModule,
                 onboardingPresentationModule,
-                onboardingDomainModule
+                onboardingDomainModule,
+                appModule
             )
         }
     }
