+++ app/src/main/java/com/amalitech/bookmeetingroom/MainViewModel.kt
@@ -1,0 +1,13 @@
+package com.amalitech.bookmeetingroom
+
+import androidx.lifecycle.ViewModel
+import com.amalitech.onboarding.preferences.OnboardingSharedPreferences
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+
+class MainViewModel(
+    sharedPreferences: OnboardingSharedPreferences
+) : ViewModel() {
+    private var _showOnBoarding: MutableStateFlow<Boolean> = MutableStateFlow(sharedPreferences.loadShouldShowOnboarding())
+    val showOnBoarding = _showOnBoarding.asStateFlow()
+}
