+++ app/src/main/java/com/amalitech/bookmeetingroom/di/AppModule.kt
@@ -1,0 +1,11 @@
+package com.amalitech.bookmeetingroom.di
+
+import com.amalitech.bookmeetingroom.MainViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val appModule = module {
+    viewModel {
+        MainViewModel(get())
+    }
+}
